requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
email-validator==2.1.0
Werkzeug==2.3.7
pycryptodome==3.19.0
python-dotenv==1.0.0
bcrypt==4.0.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
pytest==7.4.2
SQLAlchemy==2.0.21
WTForms==3.0.1
